// examples/offline-demo/convex/_generated/api.d.ts
// Generated by Stackbase. Do not edit.

export type FunctionReference<
  Type extends "query" | "mutation" | "action",
  Vis extends "public" | "internal" = "public",
  Args = any,
  Returns = any,
> = {
  readonly __type: Type;
  readonly __visibility: Vis;
  readonly __args: Args;
  readonly __returns: Returns;
};

export type Api = {
  "items": {
    add: FunctionReference<"mutation", "public", { _id?: string; listId: Id<"lists">; label: string }, Id<"items">>;
    list: FunctionReference<"query", "public", { listId: Id<"lists"> }, Array<{ _id: Id<"items">; _creationTime: number; listId: Id<"lists">; label: string; done: boolean }>>;
    toggle: FunctionReference<"mutation", "public", { id: Id<"items">; done: boolean }, null>;
  };
  "lists": {
    create: FunctionReference<"mutation", "public", { _id?: string; name: string }, Id<"lists">>;
    list: FunctionReference<"query", "public", {}, Array<{ _id: Id<"lists">; _creationTime: number; name: string; locked: boolean }>>;
    lock: FunctionReference<"mutation", "public", { id: Id<"lists"> }, null>;
  };
};

export type UdfPathOf<A> = {
  [M in keyof A & string]: {
    [F in keyof A[M] & string]: `${M}:${F}`;
  }[keyof A[M] & string];
}[keyof A & string];